[WEJBHTTP-26][WFLY-12184] authentication context passed from ejb call to transaction invocation calls
The EJB invocation uses the HttpEJBReceiver which receives
authentication configuration from the EJB layer from WFLY.
In case of the transaction calls the transaction manager
does not provide the configuration to the call again.
The ejb call needs to store the auth config data
and then they could be used during txn calls.